--- open_spiel/games/shogi/shogi_common.h
@@ -1,160 +1,0 @@
-// Copyright 2026 DeepMind Technologies Limited
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#ifndef OPEN_SPIEL_GAMES_SHOGI_COMMON_H_
-#define OPEN_SPIEL_GAMES_SHOGI_COMMON_H_
-
-#include <array>
-#include <cstdint>
-#include <random>
-#include <string>
-#include <utility>
-#include <vector>
-
-#include "open_spiel/abseil-cpp/absl/random/uniform_int_distribution.h"
-
-namespace open_spiel {
-namespace shogi {
-constexpr int kBoardSize = 9;
-constexpr int kNumSquares = 81;
-constexpr int kNumBoardMoves = 81 * 81 * 2;
-constexpr int kNumDropMoves  = 7 * 81;
-}  // shogi_common
-
-namespace shogi_common {
-
-struct Offset {
-  int8_t x_offset;
-  int8_t y_offset;
-
-  bool operator==(const Offset& other) const {
-    return x_offset == other.x_offset && y_offset == other.y_offset;
-  }
-};
-
-// x corresponds to file (column / letter)
-// y corresponds to rank (row / number).
-struct Square {
-  Square& operator+=(const Offset& offset) {
-    x += offset.x_offset;
-    y += offset.y_offset;
-    return *this;
-  }
-
-  bool operator==(const Square& other) const {
-    return x == other.x && y == other.y;
-  }
-
-  bool operator!=(const Square& other) const { return !(*this == other); }
-
-  // Required by std::set.
-  bool operator<(const Square& other) const {
-    if (x != other.x) {
-      return x < other.x;
-    } else {
-      return y < other.y;
-    }
-  }
-
-  std::string ToString() const {
-    std::string s;
-    s.push_back('a' + x);
-    s.push_back('1' + y);
-    return s;
-  }
-
-	int Index() const {
-		return y * kBoardSize + x;
-	}
-
-  int8_t x;
-  int8_t y;
-};
-
-constexpr Square kInvalidSquare{-1, -1};
-
-inline std::string SquareToString(const Square& square) {
-  if (square == kInvalidSquare) {
-    return "None";
-  } else {
-    std::string s;
-    s.push_back('a' + square.x);
-    s.push_back('1' + square.y);
-    return s;
-  }
-}
-
-inline Square operator+(const Square& sq, const Offset& offset) {
-  int8_t x = sq.x + offset.x_offset;
-  int8_t y = sq.y + offset.y_offset;
-  return Square{x, y};
-}
-
-// n-dimensional array of uniform random numbers.
-// Example:
-//   ZobristTable<int, 3, 4, 5> table;
-//
-//   table[a][b][c] is a random int where a < 3, b < 4, c < 5
-//
-template <typename T, std::size_t InnerDim, std::size_t... OtherDims>
-class ZobristTable {
- public:
-  using Generator = std::mt19937_64;
-  using NestedTable = ZobristTable<T, OtherDims...>;
-
-  explicit ZobristTable(Generator::result_type seed) {
-    Generator generator(seed);
-    absl::uniform_int_distribution<Generator::result_type> dist;
-    data_.reserve(InnerDim);
-    for (std::size_t i = 0; i < InnerDim; ++i) {
-      data_.emplace_back(dist(generator));
-    }
-  }
-
-  const NestedTable& operator[](std::size_t inner_index) const {
-    return data_[inner_index];
-  }
-
- private:
-  std::vector<NestedTable> data_;
-};
-
-// 1-dimensional array of uniform random numbers.
-template <typename T, std::size_t InnerDim>
-class ZobristTable<T, InnerDim> {
- public:
-  using Generator = std::mt19937_64;
-
-  explicit ZobristTable(Generator::result_type seed) : data_(InnerDim) {
-    Generator generator(seed);
-    absl::uniform_int_distribution<T> dist;
-    for (auto& field : data_) {
-      field = dist(generator);
-    }
-  }
-
-  T operator[](std::size_t index) const { return data_[index]; }
-
- private:
-  std::vector<T> data_;
-};
-
-inline std::ostream& operator<<(std::ostream& stream, const Square& sq) {
-  return stream << SquareToString(sq);
-}
-
-}  // namespace crazyhouse_common
-}  // namespace open_spiel
-
-#endif  // OPEN_SPIEL_GAMES_SHOGI_SHOGI_COMMON_H_
